Dissolve YamlPrinter into the Yaml.Ast Showable instance
`YamlPrinter` was a public object rendering YAML, duplicating what
`Yaml.Ast is Showable` should expose via `.show`. The block-style emission
fold now lives in that instance, so the typeclass is the single route to
YAML text; the HTTP `abstractable` uses `.show.data` and the unused
streaming `emit` is dropped.

The stateful `YamlParser` stays a class but is no longer a public second
entry point — `object YamlParser` becomes `private[ypsiloid]`, and both it
and `YamlPrinter` are dropped from the soundness re-export. `Yaml.parseAll`
/ `readAll` remain as the multi-document public API. The stale
`jacinta.JsonPrinter` reference in the deleted file's comment is gone.
